requirements.txt
aws-cdk-lib
pytube
boto3
yawsso